+++ src/tmui/components/tm-time-between/tm-time-between.vue
@@ -81,12 +81,12 @@
     <view>
       <tm-time-view
         :showDetail="props.showDetail"
-        :format="props.format"
         :immediateChange="props.immediateChange"
         :start="starTtime"
         :end="endTtime"
         @change="timeChange"
         v-model="picker_focus_date"
+		v-model:model-str="picker_focus_date_str"
         :defaultValue="picker_focus_date"
       ></tm-time-view>
     </view>
@@ -218,8 +218,18 @@ const props = defineProps({
     type: Array as PropType<Array<String | Number | Date>>,
     default: () => [],
   },
+  /**
+   * 不受formart控制，始终完整显示，必须规范
+   * */
   modelValue: {
     type: Array as PropType<Array<String | Number | Date>>,
+    default: () => [],
+  },
+  /**
+   * 只对外显示输出formart的格式字符串，可能不规范
+   * */
+  modelStr: {
+    type: Array as PropType<Array<String>>,
     default: () => [],
   },
   /** 当改变日期时，是否需要同步到vmodel变量中，立即执行双向同步，如果为false日期的改变不会同步到vmodel中
@@ -236,13 +246,14 @@ const props = defineProps({
     default: true,
   },
 });
-const emits = defineEmits(["change", "update:modelValue", "confirm", "clear"]);
+const emits = defineEmits(["change", "update:modelValue", "update:modelStr", "confirm", "clear"]);
 const valuedate_start = ref("");
 const valuedate_end = ref("");
 const focusIndex = ref(-1);
 const picker_focus_date = ref("");
+const picker_focus_date_str = ref("");
 const defaultFormat = "YYYY/MM/DD HH:mm:ss";
-let tid = NaN;
+let tid:any = NaN;
 /** 开始时间 */
 const starTtime = computed(()=>DayJs(props.start).format(defaultFormat))
 /** 结束时间 */
@@ -252,6 +263,7 @@ setDate(props.defaultValue);
 watch(
   () => props.modelValue,
   () => {
+	clearTimeout(tid)
     tid = setTimeout(() => {
       setDate(props.modelValue);
     }, 100);
@@ -262,8 +274,7 @@ const valuedate_start_str = computed(() => {
   if (!valuedate_start.value) return "";
   return DayJs(valuedate_start.value).format(props.format);
 });
-//默认进入页面是选中开始的,就要赋值默认值.
-// valuedate_start.value = valuedate_start_str.value
+
 const valuedate_end_str = computed(() => {
   if (!valuedate_end.value) return "";
   return DayJs(valuedate_end.value).format(props.format);
@@ -272,105 +283,52 @@ function changeFocus(index: number) {
   focusIndex.value = index;
   let vstd = valuedate_start.value;
   let vsed = valuedate_end.value;
-
+	
   if (index == 0 && !valuedate_start.value) {
-    vstd = valuedate_end.value || DayJs().format(defaultFormat);
+	 
+    vstd =  DayJs(valuedate_end.value).isValid()?valuedate_end.value:DayJs().format(defaultFormat)
   }
   if (index == 1 && !valuedate_end.value) {
-    vsed = valuedate_start.value || DayJs().format(defaultFormat);
+    vsed = DayJs(valuedate_start.value).isValid()?valuedate_start.value:DayJs().format(defaultFormat)
   }
-
+	
   setDate([vstd, vsed]);
 }
-function formatUpdate(evt: string,format:string){
-	if(DayJs(evt).format(format) == 'Invalid Date'){
-		let time = DayJs(new Date()).format(defaultFormat);
-		let str = evt.split('');
-		let positionY = props.format.indexOf('Y');
-		if( positionY != -1){
-			time = replaceStr(time,0,evt.substring(positionY,positionY+1))
-			time = replaceStr(time,1,evt.substring(positionY+1,positionY+2))
-			time = replaceStr(time,2,evt.substring(positionY+2,positionY+3))
-			time = replaceStr(time,3,evt.substring(positionY+3,positionY+4))
-		}
-		let positionM = props.format.indexOf('M');
-		if( positionM != -1){
-			time = replaceStr(time,5,evt.substring(positionM,positionM+1))
-			time = replaceStr(time,6,evt.substring(positionM+1,positionM+2))
-		}
-		let positionD = props.format.indexOf('D');
-		if( positionD != -1){
-			time = replaceStr(time,8,evt.substring(positionD,positionD+1))
-			time = replaceStr(time,9,evt.substring(positionD+1,positionD+2))
-		}
-		let positionH = props.format.indexOf('H');
-		if( positionH != -1){
-			time = replaceStr(time,11,evt.substring(positionH,positionH+1))
-			time = replaceStr(time,12,evt.substring(positionH+1,positionH+2))
-		}
-		let positionm = props.format.indexOf('m');
-		if( positionm != -1){
-			time = replaceStr(time,14,evt.substring(positionm,positionm+1))
-			time = replaceStr(time,15,evt.substring(positionm+1,positionm+2))
-		}
-		let positions = props.format.indexOf('s');
-		if( positions != -1){
-			time = replaceStr(time,16,evt.substring(positions,positions+1))
-			time = replaceStr(time,17,evt.substring(positions+1,positions+2))
-		}
-		return time;
-	}else{
-		return DayJs(evt).format(format)
-	}
-}
-function replaceStr(str:string, index:number, char:string){
-	 const strAry = str.split('');
-	 strAry[index] = char;
-	 return strAry.join('');
-}
+
+
 function timeChange(evt: string) {
+
   if (focusIndex.value == -1) focusIndex.value = 0;
 
   if (focusIndex.value == 0) {
-    valuedate_start.value = formatUpdate(evt,defaultFormat);
+    valuedate_start.value = picker_focus_date.value;
   } else if (focusIndex.value == 1) {
-    valuedate_end.value = formatUpdate(evt,defaultFormat);
+    valuedate_end.value = picker_focus_date.value;
   }
+
   // #ifdef APP-NVUE
-  checkTimeVaild();
-  let nst = DayJs(valuedate_start.value).isValid()
-    ? DayJs(valuedate_start.value).format(props.format)
-    : "";
-  let ese = DayJs(valuedate_end.value).isValid()
-    ? DayJs(valuedate_end.value).format(props.format)
-    : "";
-  if (props.asyncModel) {
-    emits("update:modelValue", [nst, ese]);
-  }
-  nextTick(() => {
-    emits("change", [nst, ese]);
-  });
+  updateTime()
   // #endif
 
   // #ifndef APP-NVUE
   nextTick(() => {
-    checkTimeVaild();
-    let nst = DayJs(valuedate_start.value).isValid()
-      ? DayJs(valuedate_start.value).format(props.format)
-      : "";
-    let ese = DayJs(valuedate_end.value).isValid()
-      ? DayJs(valuedate_end.value).format(props.format)
-      : "";
-    if (props.asyncModel) {
-      emits("update:modelValue", [nst, ese]);
-    }
-    nextTick(() => {
-      emits("change", [nst, ese]);
-    });
+    updateTime()
   });
   // #endif
-}
+  
+  
 
+}
+function updateTime(){
+	  checkTimeVaild();
+	  if (props.asyncModel) {
+		emits("update:modelValue", [valuedate_start.value, valuedate_end.value]);
+		emits("update:modelStr", [DayJs(valuedate_start.value).format(props.format), DayJs(valuedate_end.value).format(props.format)]);
+	  }
+	  nextTick(() => {
+		emits("change", [valuedate_start.value, valuedate_end.value]);
+	  });
+}
 //检查两个时间的正确性,就是后者要大于前者.
 function checkTimeVaild() {
   if (valuedate_end.value.length == 0 || valuedate_start.value.length == 0) return;
@@ -444,40 +402,32 @@ function quickChangeClick(item: quickBtnType) {
 }
 
 function setDate(ar: Array<String | Number | Date>) {
-  if (Array.isArray(ar)) {
-    if (ar.length == 2) {
-      valuedate_start.value = DayJs(ar[0]).isValid()
-        ? DayJs(ar[0]).format(defaultFormat)
-        : "";
-      valuedate_end.value = DayJs(ar[1]).isValid()
-        ? DayJs(ar[1]).format(defaultFormat)
-        : "";
-      focusIndex.value = focusIndex.value == -1 ? 0 : focusIndex.value;
-
-      if (!DayJs(ar[focusIndex.value]).isValid()) return;
-      picker_focus_date.value = DayJs(ar[focusIndex.value]).format(defaultFormat);
-    }
+  if (Array.isArray(ar)&&ar?.length == 2) {
+    valuedate_start.value = DayJs(ar[0]).isValid()
+      ? DayJs(ar[0]).format(defaultFormat)
+      : "";
+    valuedate_end.value = DayJs(ar[1]).isValid()
+      ? DayJs(ar[1]).format(defaultFormat)
+      : "";
+    focusIndex.value = focusIndex.value == -1 ? 0 : focusIndex.value;
+    
+    if (!DayJs(ar[focusIndex.value]).isValid()) return;
+    picker_focus_date.value = DayJs(ar[focusIndex.value]).format(defaultFormat);
   }
 }
 function confirm() {
-  let nst = valuedate_start.value
-    ? DayJs(valuedate_start.value).format(props.format)
-    : "";
-  let ese = valuedate_end.value ? DayJs(valuedate_end.value).format(props.format) : "";
-  emits("update:modelValue", [nst, ese]);
-  emits("confirm", [nst, ese]);
+  emits("update:modelValue", [valuedate_start.value, valuedate_end.value]);
+  emits("update:modelStr", [DayJs(valuedate_start.value).format(props.format), DayJs(valuedate_end.value).format(props.format)]);
+  emits("confirm", [valuedate_start.value, valuedate_end.value]);
 }
 function getNowDate(): [string, string] {
-  let nst = valuedate_start.value
-    ? DayJs(valuedate_start.value).format(props.format)
-    : "";
-  let ese = valuedate_end.value ? DayJs(valuedate_end.value).format(props.format) : "";
-  emits("update:modelValue", [nst, ese]);
-  return [nst, ese];
+  updateTime();
+  return [valuedate_start.value, valuedate_end.value];
 }
 function clear() {
   valuedate_start.value = "";
   valuedate_end.value = "";
+  emits("update:modelStr", ["",""]);
   emits("clear");
 }
 defineExpose({ getNowDate });